+++ src/standalone/standalone.ts
@@ -13,11 +13,12 @@ import { ExternalHostBridgeClientManager } from "./host-bridge-client-manager"
 import { ExternalWebviewProvider } from "./ExternalWebviewProvider"
 import { WebviewProviderType } from "@/shared/webview/types"
 import { v4 as uuidv4 } from "uuid"
+import { ExternalDiffViewProvider } from "./ExternalDiffviewProvider"
 
 async function main() {
 	log("Starting standalone service...")
 
-	hostProviders.initializeHostProviders(createWebview, new ExternalHostBridgeClientManager())
+	hostProviders.initializeHostProviders(createWebview, createDiffView, new ExternalHostBridgeClientManager())
 	activate(extensionContext)
 	const controller = new Controller(extensionContext, outputChannel, postMessage, uuidv4())
 	startProtobusService(controller)
@@ -60,8 +61,11 @@ function getProtobusServiceNames(packageDefinition: { [x: string]: any }): strin
 	return protobusServiceNames
 }
 
-const createWebview = () => {
+function createWebview() {
 	return new ExternalWebviewProvider(extensionContext, outputChannel, WebviewProviderType.SIDEBAR)
+}
+function createDiffView() {
+	return new ExternalDiffViewProvider()
 }
 
 /**
